+++ CHANGELOG.md
@@ -36,6 +36,7 @@
     -   This fixes a reported bug with `lineTo` to updating correctly while using the `transformer` tag. [Issue #276](https://github.com/casual-simulation/casualos/issues/276)
 -   Fixed an issue where the `color` tag would not apply to all materials in a gltf model.
 -   Fixed an issue where gltf models with multiple materials and textures would not be properly disposed.
+-   Fixed an issue where `os.beginAudioRecording` would fail to provide audio chunks in stream mode if the mimeType was anything other than `audio/x-raw`.
 -   Fixed some potential issues with the `tempShared` and `remoteTempShared` spaces.
 
 ## V3.1.28